Creating a common abstract class for OAuth2 methods and putting Intent and WebView based parts in their own classes Motivation: Prior to this we only had a webview implementation.  WebView and Intent based OAuth2 uses a lot of the same code/logic but have a few different bits.
